8.1.1 refactoring: change name components
