+++ src/components/board-component.js
@@ -1,0 +1,9 @@
+import AbstractComponent from './abstract-component.js';
+
+export default class BoardComponent extends AbstractComponent {
+  getTemplate() {
+    return (
+      `<section class="board container"></section>`
+    );
+  }
+}
