--- src/components/board/board.js
@@ -1,9 +1,0 @@
-import AbstractComponent from '../abstract-component.js';
-
-export default class Board extends AbstractComponent {
-  getTemplate() {
-    return (
-      `<section class="board container"></section>`
-    );
-  }
-}
