+++ src/components/load-more-button-component.js
@@ -1,6 +1,6 @@
-import AbstractComponent from '../abstract-component.js';
+import AbstractComponent from './abstract-component.js';
 
-export default class LoadMoreButton extends AbstractComponent {
+export default class LoadMoreButtonComponent extends AbstractComponent {
   getTemplate() {
     return (
       `<button class="load-more" type="button">load more</button>`
