--- src/components/task/tasks-container.js
@@ -1,9 +1,0 @@
-import AbstractComponent from '../abstract-component.js';
-
-export default class TasksContainer extends AbstractComponent {
-  getTemplate() {
-    return (
-      `<div class="board__tasks"></div>`
-    );
-  }
-}
