+++ src/components/tasks-container-component.js
@@ -1,0 +1,9 @@
+import AbstractComponent from './abstract-component.js';
+
+export default class TasksContainerComponent extends AbstractComponent {
+  getTemplate() {
+    return (
+      `<div class="board__tasks"></div>`
+    );
+  }
+}
