+++ src/main.js
@@ -1,17 +1,17 @@
-import SiteMenu, {MenuItem} from './components/menu/site-menu.js';
-import FilterController from './controllers/filter.js';
-import Board from './components/board/board.js';
+import BoardComponent from './components/board-component.js';
+import SiteMenuComponent, {MenuItem} from './components/site-menu-component.js';
+import StatisticsComponent from './components/statistics-component.js';
 import BoardController from './controllers/board.js';
-import Statistics from './components/statistics/statistics.js';
+import FilterController from './controllers/filter.js';
 import TasksModel from './models/tasks.js';
-import {generateTasks} from './mock/task.js';
 import {render} from './utils/render.js';
+import {generateTasks} from './mock/task.js';
 
 const TASK_COUNT = 20;
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
-const siteMenuComponent = new SiteMenu();
+const siteMenuComponent = new SiteMenuComponent();
 render(siteHeaderElement, siteMenuComponent);
 
 const tasks = generateTasks(TASK_COUNT);
@@ -21,7 +21,7 @@ tasksModel.setTasks(tasks);
 const filterController = new FilterController(siteMainElement, tasksModel);
 filterController.render();
 
-const boardComponent = new Board();
+const boardComponent = new BoardComponent();
 render(siteMainElement, boardComponent);
 const boardController = new BoardController(boardComponent, tasksModel);
 boardController.render();
@@ -32,7 +32,7 @@ const dateFrom = (() => {
   d.setDate(d.getDate() - 7);
   return d;
 })();
-const statisticsComponent = new Statistics(tasksModel, dateFrom, dateTo);
+const statisticsComponent = new StatisticsComponent(tasksModel, dateFrom, dateTo);
 render(siteMainElement, statisticsComponent);
 statisticsComponent.hide();
